Remove handlebars updates since we are using twigJS
