+++ styleguide/source/assets/js/modules/pagination.js
@@ -5,6 +5,47 @@ export default function (window,document,$,undefined) {
     return;
   }
 
+  // {{compare unicorns ponies operator="<"}}
+  // 	I knew it, unicorns are just low-quality ponies!
+  // {{/compare}}
+  //
+  // (defaults to == if operator omitted)
+  //
+  // {{equal unicorns ponies }}
+  // 	That's amazing, unicorns are actually undercover ponies
+  // {{/equal}}
+  // (from http://doginthehat.com.au/2012/02/comparison-block-helper-for-handlebars-templates/)
+  Handlebars.registerHelper('compare', function(lvalue, rvalue, options) {
+
+    if (arguments.length < 3)
+      throw new Error("Handlerbars Helper 'compare' needs 2 parameters");
+
+    let operator = options.hash.operator || "==";
+
+    let operators = {
+      '==':		function(l,r) { return l == r; },
+      '===':	function(l,r) { return l === r; },
+      '!=':		function(l,r) { return l != r; },
+      '<':		function(l,r) { return l < r; },
+      '>':		function(l,r) { return l > r; },
+      '<=':		function(l,r) { return l <= r; },
+      '>=':		function(l,r) { return l >= r; },
+      'typeof':	function(l,r) { return typeof l == r; }
+    };
+
+    if (!operators[operator])
+      throw new Error("Handlerbars Helper 'compare' doesn't know the operator "+operator);
+
+    let result = operators[operator](lvalue,rvalue);
+
+    if( result ) {
+      return options.fn(this);
+    } else {
+      return options.inverse(this);
+    }
+
+  });
+
   // Set up global component config
   let compiledTemplate = getTemplate('pagination'),
     prevButton = '.js-pagination-prev',
